When installing Lakebridge, detect and upgrade installed transpilers if necessary.
In addition to this, a lot of the installer code specific to the transpilers has been moved to the transpilers module.
