+++ tests/resources/lsp_transpiler/test-lsp-server.log
@@ -1,0 +1,43 @@
+INFO:pygls.feature_manager:Registered builtin feature exit
+INFO:pygls.feature_manager:Registered builtin feature initialize
+INFO:pygls.feature_manager:Registered builtin feature initialized
+INFO:pygls.feature_manager:Registered builtin feature notebookDocument/didChange
+INFO:pygls.feature_manager:Registered builtin feature notebookDocument/didClose
+INFO:pygls.feature_manager:Registered builtin feature notebookDocument/didOpen
+INFO:pygls.feature_manager:Registered builtin feature $/setTrace
+INFO:pygls.feature_manager:Registered builtin feature shutdown
+INFO:pygls.feature_manager:Registered builtin feature textDocument/didChange
+INFO:pygls.feature_manager:Registered builtin feature textDocument/didClose
+INFO:pygls.feature_manager:Registered builtin feature textDocument/didOpen
+INFO:pygls.feature_manager:Registered builtin feature window/workDoneProgress/cancel
+INFO:pygls.feature_manager:Registered builtin feature workspace/didChangeWorkspaceFolders
+INFO:pygls.feature_manager:Registered builtin feature workspace/executeCommand
+INFO:pygls.feature_manager:Registered "initialize" with options "None"
+INFO:pygls.feature_manager:Registered "textDocument/didOpen" with options "None"
+INFO:pygls.feature_manager:Registered "textDocument/didClose" with options "None"
+INFO:pygls.feature_manager:Registered "document/transpileToDatabricks" with options "None"
+DEBUG:__main__:SOME_ENV=abc
+DEBUG:__main__:sys.args=['lsp_server.py', '--stuff=12', '--log_level=INFO']
+INFO:pygls.server:Starting async IO server
+DEBUG:asyncio:Using selector: KqueueSelector
+DEBUG:pygls.server:Content length: 439
+DEBUG:pygls.protocol.json_rpc:Request 'initialize' received
+INFO:pygls.protocol.language_server:Language server initialized InitializeParams(capabilities=ClientCapabilities(workspace=None, text_document=None, notebook_document=None, window=None, general=None, experimental=None), process_id=76534, client_info=ClientInfo(name='lakebridge', version='0.10.8+320250829113309'), locale=None, root_path=None, root_uri='file:///Users/andrew.snare/dev/lakebridge', initialization_options={'remorph': {'source-dialect': 'source_dialect'}, 'options': None, 'custom': {'whatever': 'xyz'}}, trace=None, work_done_token=None, workspace_folders=None)
+DEBUG:__main__:client-info=lakebridge/0.10.8+320250829113309
+DEBUG:__main__:client-process-id=76534
+DEBUG:__main__:dialect=source_dialect
+DEBUG:__main__:whatever=xyz
+DEBUG:__main__:experimental=None
+DEBUG:pygls.protocol.json_rpc:Sending request with id "2b93fa95-62ed-4f9e-89f0-7f51ac92490e": client/registerCapability RegistrationParams(registrations=[Registration(id='2e1eb1ac-ab49-4743-803d-bfee723c52cb', method='document/transpileToDatabricks', register_options=None)])
+INFO:pygls.protocol.json_rpc:Sending data: {"id": "2b93fa95-62ed-4f9e-89f0-7f51ac92490e", "params": {"registrations": [{"id": "2e1eb1ac-ab49-4743-803d-bfee723c52cb", "method": "document/transpileToDatabricks"}]}, "method": "client/registerCapability", "jsonrpc": "2.0"}
+DEBUG:pygls.server:Content length: 80
+DEBUG:pygls.protocol.json_rpc:Response message received.
+DEBUG:pygls.protocol.json_rpc:Received result for message "2b93fa95-62ed-4f9e-89f0-7f51ac92490e": None
+DEBUG:pygls.protocol.language_server:Server capabilities: {"positionEncoding": "utf-16", "textDocumentSync": {"openClose": true, "change": 2, "save": false}, "executeCommandProvider": {"commands": []}, "workspace": {"workspaceFolders": {"supported": true, "changeNotifications": true}, "fileOperations": {}}}
+INFO:pygls.protocol.json_rpc:Sending data: {"id": "28898a92-db71-4e84-a7f1-93826771679d", "jsonrpc": "2.0", "result": {"capabilities": {"positionEncoding": "utf-16", "textDocumentSync": {"openClose": true, "change": 2, "save": false}, "executeCommandProvider": {"commands": []}, "workspace": {"workspaceFolders": {"supported": true, "changeNotifications": true}, "fileOperations": {}}}, "serverInfo": {"name": "test-lsp-server", "version": "v0.1"}}}
+DEBUG:pygls.server:Content length: 86
+DEBUG:pygls.protocol.json_rpc:Request 'shutdown' received
+INFO:pygls.protocol.json_rpc:Sending data: {"id": "ab63a36e-0a82-474c-8154-0f5b5618d713", "jsonrpc": "2.0", "result": null}
+DEBUG:pygls.server:Content length: 36
+DEBUG:pygls.protocol.json_rpc:Notification 'exit' received
+INFO:pygls.server:Shutting down the server
