+++ tests/unit/test_install.py
@@ -1,14 +1,10 @@
-import datetime as dt
-import json
 import logging
-import os
-import shutil
 from collections.abc import Callable, Generator, Sequence
 from pathlib import Path
 from unittest.mock import create_autospec, patch
 
 import pytest
-from databricks.labs.blueprint.installation import MockInstallation
+from databricks.labs.blueprint.installation import JsonObject, MockInstallation
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.service import iam
 from databricks.labs.blueprint.tui import MockPrompts
@@ -25,8 +21,9 @@ from databricks.labs.lakebridge.config import (
 from databricks.labs.lakebridge.contexts.application import ApplicationContext
 from databricks.labs.lakebridge.deployment.configurator import ResourceConfigurator
 from databricks.labs.lakebridge.deployment.installation import WorkspaceInstallation
-from databricks.labs.lakebridge.install import WorkspaceInstaller, TranspilerInstaller
+from databricks.labs.lakebridge.install import WorkspaceInstaller
 from databricks.labs.lakebridge.reconcile.constants import ReconSourceType, ReconReportType
+from databricks.labs.lakebridge.transpiler.installers import TranspilerInstaller
 from databricks.labs.lakebridge.transpiler.repository import TranspilerRepository
 
 from tests.unit.conftest import path_to_resource
@@ -57,18 +54,14 @@ def ws_installer() -> Generator[Callable[..., WorkspaceInstaller], None, None]:
 
     class TestWorkspaceInstaller(WorkspaceInstaller):
         def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
             # Ensure that the transpiler repository is mocked for unit tests instead of being the real thing.
-            if self._transpiler_repository == TranspilerRepository.user_home():
-                self._transpiler_repository = create_autospec(TranspilerRepository)
+            if "transpiler_repository" not in kwargs:
+                kwargs["transpiler_repository"] = create_autospec(TranspilerRepository)
+            # In these unit tests we have no transpilers to install by default.
+            if "transpiler_installers" not in kwargs:
+                kwargs["transpiler_installers"] = ()
 
-        def install_bladebridge(self, artifact: Path | None = None) -> None:
-            # Ensure that unit tests do not attempt to install BladeBridge
-            pass
-
-        def install_morpheus(self, artifact: Path | None = None) -> None:
-            # Ensure that unit tests do not attempt to install Morpheus
-            pass
+            super().__init__(*args, **kwargs)
 
         def _all_installed_dialects(self):
             return ALL_INSTALLED_DIALECTS_NO_LATER
@@ -292,12 +285,7 @@ def test_configure_transpile_no_existing_installation(
     )
 
 
-@patch("databricks.labs.lakebridge.install.WorkspaceInstaller.install_bladebridge")
-@patch("databricks.labs.lakebridge.install.WorkspaceInstaller.install_morpheus")
-def test_configure_transpile_installation_no_override(mock_install_morpheus, mock_install_bladebridge, ws):
-    mock_install_bladebridge.return_value = None
-    mock_install_morpheus.return_value = None
-
+def test_configure_transpile_installation_no_override(ws: WorkspaceClient) -> None:
     prompts = MockPrompts(
         {
             r"Do you want to override the existing installation?": "no",
@@ -1358,113 +1346,9 @@ def test_runs_and_stores_choice_config_option(
     )
 
 
-def test_store_product_state(tmp_path) -> None:
-    """Verify the product state is stored after installing is correct."""
-
-    class MockTranspilerInstaller(TranspilerInstaller):
-        @classmethod
-        def store_product_state(cls, product_path: Path, version: str) -> None:
-            cls._store_product_state(product_path, version)
-
-    # Store the product state, capturing the time before and after so we can verify the timestamp it puts in there.
-    before = dt.datetime.now(tz=dt.timezone.utc)
-    MockTranspilerInstaller.store_product_state(tmp_path, "1.2.3")
-    after = dt.datetime.now(tz=dt.timezone.utc)
-
-    # Load the state that was just stored.
-    with (tmp_path / "state" / "version.json").open("r", encoding="utf-8") as f:
-        stored_state = json.load(f)
-
-    # Verify the timestamp first.
-    stored_date = stored_state["date"]
-    parsed_date = dt.datetime.fromisoformat(stored_date)
-    assert parsed_date.tzinfo is not None, "Stored date should be timezone-aware."
-    assert before <= parsed_date <= after, f"Stored date {stored_date} is not within the expected range."
-
-    # Verify the rest, now that we've checked the timestamp.
-    expected_state = {
-        "version": "v1.2.3",
-        "date": stored_date,
-    }
-    assert stored_state == expected_state
-
-
-@pytest.fixture
-def no_java(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Ensure that (temporarily) no 'java' binary can be found in the environment."""
-    found_java = shutil.which("java")
-    while found_java is not None:
-        # Java is installed, so we need to figure out how to remove it from the path.
-        # (We loop here to handle cases where multiple java binaries are available via the PATH.)
-        java_directory = Path(found_java).parent
-        search_path = os.environ.get("PATH", os.defpath).split(os.pathsep)
-        updated_path = os.pathsep.join(p for p in search_path if p and Path(p) != java_directory)
-        assert (
-            search_path != updated_path
-        ), f"Did not find {java_directory} in {search_path}, but 'java' was found at {found_java}."
-
-        # Set the modified PATH without the directory where 'java' was found.
-        monkeypatch.setenv("PATH", os.pathsep.join(updated_path))
-
-        # Check again if 'java' is still found.
-        found_java = shutil.which("java")
-
-
-def test_java_version_with_java_missing(no_java: None) -> None:
-    """Verify the Java version check handles Java missing entirely."""
-    expected_missing = WorkspaceInstaller.find_java()
-    assert expected_missing is None
-
-
-class FriendOfWorkspaceInstaller(WorkspaceInstaller):
-    """A friend class to access protected methods for testing purposes."""
-
-    @classmethod
-    def parse_java_version(cls, output: str) -> tuple[int, int, int, int] | None:
-        return cls._parse_java_version(output)
-
-
-@pytest.mark.parametrize(
-    ("version", "expected"),
-    (
-        # Real examples.
-        pytest.param("1.8.0_452", None, id="1.8.0_452"),
-        pytest.param("11.0.27", (11, 0, 27, 0), id="11.0.27"),
-        pytest.param("17.0.15", (17, 0, 15, 0), id="17.0.15"),
-        pytest.param("21.0.7", (21, 0, 7, 0), id="21.0.7"),
-        pytest.param("24.0.1", (24, 0, 1, 0), id="24.0.1"),
-        # All digits.
-        pytest.param("1.2.3.4", (1, 2, 3, 4), id="1.2.3.4"),
-        # Trailing zeros can be omitted.
-        pytest.param("1.2.3", (1, 2, 3, 0), id="1.2.3"),
-        pytest.param("1.2", (1, 2, 0, 0), id="1.2"),
-        pytest.param("1", (1, 0, 0, 0), id="1"),
-        # Another edge case.
-        pytest.param("", None, id="empty string"),
-    ),
-)
-def test_java_version_parse(version: str, expected: tuple[int, int, int, int] | None) -> None:
-    """Verify that the Java version parsing works correctly."""
-    # Format reference: https://docs.oracle.com/en/java/javase/11/install/version-string-format.html
-    version_output = f'openjdk version "{version}" 2025-06-19'
-    parsed = FriendOfWorkspaceInstaller.parse_java_version(version_output)
-    assert parsed == expected
-
-
-def test_java_version_parse_missing() -> None:
-    """Verify that we return None when the version is missing."""
-    version_output = "Nothing in here that looks like a version."
-    parsed = FriendOfWorkspaceInstaller.parse_java_version(version_output)
-    assert parsed is None
-
-
-@pytest.mark.parametrize(("installed_transpilers", "is_upgrade"), (({"foo", "bar"}, True), ({}, False)))
-def test_installer_upgrade_detection(
-    ws_installer: Callable[..., WorkspaceInstaller],
-    ws: WorkspaceClient,
-    installed_transpilers: set[str],
-    is_upgrade: bool,
-    caplog,
+@pytest.mark.parametrize(("installed_transpilers",), (({"foo", "bar"},), ({},)))
+def test_installer_detects_installed_transpilers(
+    ws_installer: Callable[..., WorkspaceInstaller], ws: WorkspaceClient, installed_transpilers: set[str], caplog
 ) -> None:
     """Check detection of whether transpilers are already installed or not."""
     mock_repository = create_autospec(TranspilerRepository)
@@ -1483,9 +1367,154 @@ def test_installer_upgrade_detection(
     )
 
     with caplog.at_level(logging.INFO):
-        result = installer.has_installed_transpilers()
+        installer.upgrade_installed_transpilers()
 
-    assert result == is_upgrade
-    if is_upgrade:
+    if installed_transpilers:
         info_messages = [log.message for log in caplog.records if log.levelno == logging.INFO]
         assert f"Detected installed transpilers: {sorted(installed_transpilers)}" in info_messages
+
+
+def test_installer_upgrade_installed_transpilers(
+    ws_installer: Callable[..., WorkspaceInstaller], ws: WorkspaceClient
+) -> None:
+    """Check that during install we attempt to upgrade any known transpilers that are already installed."""
+
+    # The setup here is:
+    #   - 'foo' and 'bar' are installed already.
+    #   - 'bar' and 'baz' are known transpilers.
+    # It should therefore try to install/upgrade bar but _not_ baz.
+
+    mock_repository = create_autospec(TranspilerRepository)
+    mock_repository.all_transpiler_names.return_value = {"foo", "bar"}
+    ctx = ApplicationContext(ws).replace(
+        product_info=ProductInfo.for_testing(LakebridgeConfiguration),
+        prompts=(MockPrompts({r"Do you want to override the existing installation?": "no"})),
+        installation=MockInstallation({"config.yml": {"version": 3}}),
+    )
+
+    class MockTranspilerInstaller(TranspilerInstaller):
+        def __init__(self, repository: TranspilerRepository, name: str) -> None:
+            super().__init__(repository)
+            self._name = name
+            self.installed = False
+
+        def can_install(self, artifact: Path) -> bool:
+            return False
+
+        @property
+        def name(self) -> str:
+            return self._name
+
+        def install(self, artifact: Path | None = None) -> bool:
+            self.installed = True
+            return True
+
+        def mock_factory(self, repository: TranspilerRepository) -> TranspilerInstaller:
+            assert repository is self._transpiler_repository
+            return self
+
+    bar_installer = MockTranspilerInstaller(mock_repository, "bar")
+    baz_installer = MockTranspilerInstaller(mock_repository, "baz")
+
+    installer = ws_installer(
+        ctx.workspace_client,
+        ctx.prompts,
+        ctx.installation,
+        ctx.install_state,
+        ctx.product_info,
+        ctx.resource_configurator,
+        ctx.workspace_installation,
+        transpiler_repository=mock_repository,
+        transpiler_installers=(baz_installer.mock_factory, bar_installer.mock_factory),
+    )
+    upgraded_something = installer.upgrade_installed_transpilers()
+
+    assert upgraded_something, "Expected to upgrade at least one transpiler"
+    assert bar_installer.installed, "Expected 'bar' transpiler to be upgraded"
+    assert not baz_installer.installed, "Did not expect 'baz' transpiler to be upgraded"
+
+
+@pytest.mark.parametrize("test_upgrade", (True, False))
+def test_installer_upgrade_configure_if_changed(
+    ws_installer: Callable[..., WorkspaceInstaller],
+    ws: WorkspaceClient,
+    test_upgrade: bool,
+) -> None:
+    """Check that during an upgrade we reconfigure if any transpiler upgrades occurred."""
+
+    # The setup here is:
+    #   - 'foo' is the installed transpiler, and we will attempt to upgrade it.
+    #   - parameterized on whether the upgrade is necessary.
+    #   - if it was, we expect the prompt-adjusted configuration to be returned.
+
+    mock_repository = create_autospec(TranspilerRepository)
+    mock_repository.all_transpiler_names.return_value = {"foo"}
+    prior_source_dialect = "original_dialect"
+    prior_configuration: JsonObject = {
+        "version": 3,
+        "transpiler_config_path": PATH_TO_TRANSPILER_CONFIG,
+        "source_dialect": prior_source_dialect,
+        "input_source": "sf_queries",
+        "output_folder": "out_dir",
+        "error_file_path": "error_log.log",
+        "skip_validation": True,
+        "catalog_name": "remorph",
+        "schema_name": "transpiler",
+    }
+    mock_installation = MockInstallation({"config.yml": prior_configuration})
+    ctx = ApplicationContext(ws).replace(
+        product_info=ProductInfo.for_testing(LakebridgeConfiguration),
+        prompts=MockPrompts(
+            {
+                r"Do you want to override the existing installation?": "yes",
+                r"Select the source dialect": "2",
+                r"Select the transpiler": "1",
+                r"Enter .*": "/tmp/updated",
+                r"Would you like to validate.*": "no",
+                r"Open config file .* in the browser?": "no",
+            }
+        ),
+        installation=mock_installation,
+    )
+
+    class MockTranspilerInstaller(TranspilerInstaller):
+        def __init__(self, repository: TranspilerRepository) -> None:
+            super().__init__(repository)
+            self.installed = False
+
+        def can_install(self, artifact: Path) -> bool:
+            return False
+
+        @property
+        def name(self) -> str:
+            return "foo"
+
+        def install(self, artifact: Path | None = None) -> bool:
+            self.installed = True
+            return test_upgrade
+
+    installer = ws_installer(
+        ctx.workspace_client,
+        ctx.prompts,
+        ctx.installation,
+        ctx.install_state,
+        ctx.product_info,
+        ctx.resource_configurator,
+        ctx.workspace_installation,
+        transpiler_repository=mock_repository,
+        transpiler_installers=(MockTranspilerInstaller,),
+    )
+    upgraded_something = installer.upgrade_installed_transpilers()
+
+    assert upgraded_something == test_upgrade
+    if test_upgrade:
+        expected_configuration = {
+            **prior_configuration,
+            "version": 3,
+            # These were updated by the configuration.
+            "source_dialect": "tsql",
+            "input_source": "/tmp/updated",
+            "output_folder": "/tmp/updated",
+            "error_file_path": "/tmp/updated",
+        }
+        mock_installation.assert_file_written("config.yml", expected_configuration)
